skip S3 integration tests when SeaweedFS is unreachable, except on CI
Five tests round-trip the real SeaweedFS S3 endpoint and failed locally when
the service wasn't running. Extract the reachability probe from
s3-integration.test.ts into tests/s3.helpers.ts and guard the S3-dependent
tests with test.skipIf(!s3Available). On CI a missing endpoint still throws —
an absent service is a setup bug, not a skip condition.
